🖋️Rename: 패키지명 변경 및 기타 수정
