+++ src/main/java/com/bombombom/devs/global/audit/BaseEntity.java
@@ -1,19 +1,19 @@
-package com.bombombom.devs.global.domain;
+package com.bombombom.devs.global.audit;
 
-import com.bombombom.devs.global.domain.listener.Auditable;
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
 import lombok.Data;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.time.LocalDateTime;
+
 @Data
 @MappedSuperclass
 @EntityListeners(value = AuditingEntityListener.class)
-public class BaseEntity implements Auditable {
+public class BaseEntity {
     @CreatedDate
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
