--- src/main/java/com/bombombom/devs/global/domain/listener/Auditable.java
@@ -1,11 +1,0 @@
-package com.bombombom.devs.global.domain.listener;
-
-import java.time.LocalDateTime;
-
-public interface Auditable {
-    LocalDateTime getCreatedAt();
-    LocalDateTime getUpdatedAt();
-
-    void setCreatedAt(LocalDateTime createdAt);
-    void setUpdatedAt(LocalDateTime updatedAt);
-}
